Add notifications for user logout/login
This involved a few things:

Add a list client-side and server-side of the logged in users and the count
that each logged in user was logged in.

Tell the view to do something when the last instance of a user loggs out or the
first instance of a user loggs in.