Add [at]mentions
Now users can mention each other which appears as a notification and rings the
message bell.

@all -- messages all the users logged in

@username -- displays the message publically to everyone, but on the computer of
the user who is logged in as username, a ding is played and the entire row of
new content is highlighted.

all mentions are bolded

as a side effect, the notifications style is now more rigidly defined (so that
it can override a colored class).

Fixes #29