Suppress sounds and notifications on load
fixes #27
When loading the initial messages, don't play sounds or display notifications
for unread messages. This includes not making noises when a person is mentioned,
and not making
noises if the user switched away from the window before the messages loaded.
Also, don't display "unread mesages" for these initial messages. All of the
messages are unread.

I'm pretty sure this won't be a problem in a race condition for receiving
new messages. Even so, it's not such a big deal. The drawback is that the new
message is treated as an old message. But that's not much of a drawback.. yet.